// main-backend/frontend/src/generated/model/account-edit-request-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Data of an account that needs to be changed.
 */
export interface AccountEditRequestDto { 
    /**
     * New email for specified account.
     */
    email?: string;
    /**
     * New first name for specified account.
     */
    firstName?: string;
    /**
     * New last name for specified account.
     */
    lastName?: string;
    /**
     * New password for specified account.
     */
    newPassword?: string;
    /**
     * Current password to confirm password and email changes.
     */
    oldPassword?: string;
}